+++ quanmiaolightapp-20240801/alibabacloud_quanmiaolightapp20240801/models/_model_usage.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class ModelUsage(DaraModel):
+    def __init__(
+        self,
+        input_tokens: int = None,
+        output_tokens: int = None,
+        total_tokens: int = None,
+    ):
+        self.input_tokens = input_tokens
+        self.output_tokens = output_tokens
+        self.total_tokens = total_tokens
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.input_tokens is not None:
+            result['inputTokens'] = self.input_tokens
+
+        if self.output_tokens is not None:
+            result['outputTokens'] = self.output_tokens
+
+        if self.total_tokens is not None:
+            result['totalTokens'] = self.total_tokens
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('inputTokens') is not None:
+            self.input_tokens = m.get('inputTokens')
+
+        if m.get('outputTokens') is not None:
+            self.output_tokens = m.get('outputTokens')
+
+        if m.get('totalTokens') is not None:
+            self.total_tokens = m.get('totalTokens')
+
+        return self
+